+++ CHANGELOG.md
@@ -2122,4 +2122,11 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 ###  291_jquery_73 ~ [Issue 291](https://github.com/bradyhouse/house/issues/291)
   
   1.  Added [fiddles/jquery/fiddle-0073-FrameCookies](fiddles/jquery/fiddle-0073-FrameCookies)
-* Added [fiddles/jquery/fidde-0074-AgGridEnterprise](fiddles/jquery/fidde-0074-AgGridEnterprise)
+  
+  
+###  294_jquery_74 ~ [Issue 293](https://github.com/bradyhouse/house/issues/293)
+
+* **jquery**
+  
+  1.  Added [fiddles/jquery/fiddle-0074-AgGridEnterprise](fiddles/jquery/fiddle-0074-AgGridEnterprise)
+